LayoutTests/fast/scrolling/scroll-position-on-reload-rtl-expected.txt

PASS: scrollLeft after reload is -200
